Using {{ formfield field.name }} gives better UI
The downside is this needs more verbose and repetitive
player field definitions (apparently unavoidable in the
oTree Studio interface).

The upside is field validation when use clicks 'Next',
making this much more pleasant to use with 10 fields.
And this avoids having to keep the raw HTML in sync
with any future oTree changes.

Note using {{ formfield field.name label="" }} to supress
the form label which I have put in the first column.