Use group_by_arrival_time; drop alternating framing
Using group_by_arrival_time alters how the groups are
allocated, and I was unable to get the allocation of
alternating framing to still work. i.e. First 6 to start
the volunteering game are farmers, next 6 are community
centre users, etc.

We were planning to use two parallel sessions for the two
framings anyway, so simplest to drop the code to alternate
the groups' framing and can now just use the session
configuration directly & drop the extra participant field.