Use horizontal radio buttons in lottery
Also include the word points to be more explicit
(at the cost of needing a wider screen #8).

This is to address user confusion as per #22.
The layout looks like A and B in separate columns,
but is dynamic - increasing the font size and/or
reducing the window width will place the A and B
radio options one under the other (as before).