More explicit lottery layout
This is closer to the original protocol design from
Simone and Laure, and abandons my more compact version
which did not fare as well with user testing.

Closes #22.

This has the downside of needing much more vertical
space (so more scrolling, especially on smaller
screens or if using larger fonts), cross reference #8
(potential usage on a narrow phone screen).